pyproject.toml
[tool.poetry]
name = "ps_product_takehome"
version = "0.1.0"
description = "Generic environment for classifier takehome"
authors = ["jason-whittle"]

[tool.poetry.dependencies]
python = "3.9.16"
snowflake-connector-python = {extras = ["pandas"], version = "^3.7.0"}
snowflake-sqlalchemy = "^1.5.1"
boto3 = "^1.34.38"
pandas = "^2.2.2"
ipykernel = "^6.29.4"
ipywidgets = "^8.1.2"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
numpy = "^1.26.4"
statsmodels = "^0.14.2"
scikit-learn = "^1.4.2"
pyyaml = "^6.0.1"
xgboost = "^2.0.3"
openpyxl = "^3.1.2"
jinja2 = "^3.1.3"
imbalanced-learn = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
